requirements.txt
cycler==0.11.0
epynet==1.1
fonttools==4.33.3
kafka-python==2.0.2
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scipy==1.8.1
six==1.16.0
wntr==0.4.1
